integrated redis and containerized it again

+++ src/main/java/com/app/ecom_app/service/ProductService.java
@@ -1,21 +1,17 @@
 package com.app.ecom_app.service;
 
-import com.app.ecom_app.dto.OrderItemDTO;
-import com.app.ecom_app.dto.OrderResponse;
 import com.app.ecom_app.dto.ProductRequest;
 import com.app.ecom_app.dto.ProductResponse;
 import com.app.ecom_app.model.CartItem;
-import com.app.ecom_app.model.Order;
-import com.app.ecom_app.model.OrderItem;
 import com.app.ecom_app.model.Product;
 import com.app.ecom_app.repository.ProductRepo;
+import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,9 +22,12 @@ public class ProductService {
 
     private final CartService cartService;
 
+    private final RedisService redisService;
+
     public ProductResponse createProduct(ProductRequest productRequest) {
         Product savedProduct = mapProductRequestToProduct(productRequest);//convert from req->product
         productRepo.save(savedProduct);//save product
+        redisService.deleteAll();//re-cache
         return mapProductToResponse(savedProduct);//convert product->res
     }
     public Product mapProductRequestToProduct(ProductRequest productRequest) {
@@ -55,12 +54,31 @@ public class ProductService {
     }
 
     public boolean fetchProductById(String id) {
-        return productRepo.findById(id).isPresent();
+
+        String key = "product:" + id;
+        Product cachedProduct = redisService.get(key,Product.class);
+        if (cachedProduct != null) {
+            return true;
+        }
+
+        Optional<Product> productOptional = productRepo.findById(id);
+        if (productOptional.isPresent()) {
+            redisService.set(key,productOptional.get(),300L);
+            return true;
+        }
+
+        return false;
     }
 
     public ProductResponse updateProduct(String id, ProductRequest product) {
-        if(!productRepo.findById(id).isPresent()) return null;
-        Product savedProduct = productRepo.findById(id).get();
+
+        Optional<Product> productOptional = productRepo.findById(id);
+        if (productOptional.isEmpty()) {
+            return null;
+        }
+
+        Product savedProduct = productOptional.get();
+
         savedProduct.setName(product.getName());
         savedProduct.setCategory(product.getCategory());
         savedProduct.setImageUrl(product.getImageUrl());
@@ -68,43 +86,78 @@ public class ProductService {
         savedProduct.setQuantity(product.getQuantity());
         savedProduct.setDescription(product.getDescription());
         productRepo.save(savedProduct);
+        redisService.delete("product:" + id);
         return mapProductToResponse(savedProduct);
 
     }
 
     public List<ProductResponse> getAllProducts() {
 //        the naming convention automatically causes the repo to define the required query
-        return productRepo.findByActiveTrue().stream()
+
+        String key = "product:all";
+        List<ProductResponse> cachedProducts = redisService.get(key, new TypeReference<List<ProductResponse>>() {});
+
+        if (cachedProducts != null) return cachedProducts;
+
+        List<ProductResponse> products = productRepo.findByActiveTrue().stream()
                 .map(this::mapProductToResponse)
                 .collect(Collectors.toList());
+
+        redisService.set(key,products, 300L);
+        return products;
     }
 
     public void deActivateProduct(String id) {
-//        productRepo.deleteById(id);
-//        we won't remove the product from the db rather we shall it in-active
-        Product savedProduct = productRepo.findById(id).get();
-        savedProduct.setActive(false);
-        productRepo.save(savedProduct);
+
+        Product product = productRepo.findById(id).get();
+        if(product==null){
+            return;
+        }
+        product.setActive(false);
+        productRepo.save(product);
+        redisService.delete("product:" + id);
     }
 
     public boolean isProductActive(String id) {
-        return productRepo.findById(id).get().getActive();
+        return productRepo.findById(id)
+                .map(Product::getActive)
+                .orElse(false);
     }
 
     public List<ProductResponse> searchProducts(String keyword) {
-        return productRepo.searchProducts(keyword).stream()
+
+        String key = "product:search:" + keyword.toLowerCase().trim();
+
+        List<ProductResponse> cachedProductList = redisService.get(key, new TypeReference<List<ProductResponse>>() {});
+
+        if (cachedProductList != null) {
+            return cachedProductList;
+        }
+        List<ProductResponse> products = productRepo.searchProducts(keyword).stream()
                 .map(this::mapProductToResponse)
                 .collect(Collectors.toList());
+
+        redisService.set(key,products,300L);
+
+        return products;
     }
 
     public void deleteProduct(String id) {
-        productRepo.deleteById(id);
+       Optional<Product> product =  productRepo.findById(id);
+       if(product.isEmpty()){
+           return;
+       }
+       productRepo.deleteById(id);
+       redisService.delete("product:" + id);
     }
 
     public void activateProduct(String id) {
-        Product savedProduct = productRepo.findById(id).get();
-        savedProduct.setActive(true);
-        productRepo.save(savedProduct);
+       Product product = productRepo.findById(id).orElse(null);
+       if(product == null){ return;}
+
+        product.setActive(true);
+        productRepo.save(product);
+        redisService.delete("product:"+id);
     }
 
     public void updateStockQuantity(String userId) {
@@ -113,6 +166,7 @@ public class ProductService {
             productRepo.findById(cartItem.getProduct().getId()).ifPresent(product -> {
                 product.setQuantity(product.getQuantity().subtract(cartItem.getQuantityOnHand()));
                 productRepo.save(product);
+                redisService.delete("product:" + product.getId());
             });
         }
     }
